+++ src/main/java/com/by/repository/MenuRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by yagamai on 15-12-11.
  */
-public interface MenuRepository extends CrudRepository<Menu, Long> {
+public interface MenuRepository extends CrudRepository<Menu, Integer> {
 }
